Improve NAME-CONFLICT for USE-PACKAGE
After checking for a package lock, and for duplicates in the package
list (and redundant entries with the existing package use list),
collect all external symbols from all to-be-used packages of the same
name.  For each of those names, check whether the operation will
result in a name conflict either with an existing accessible symbol or
between the packages, and if so call NAME-CONFLICT for resolution,
before actually performing the low-level package-table manipulation
for using the packages.

Make NAME-CONFLICT for USE-PACKAGE take as a datum not a single
symbol, but all the symbols from the used packages of the same name.
Any already-accessible symbol, if different, should be passed in as an
element of the SYMBOLS &rest list.

Make the pairwise restarts KEEP-OLD and TAKE-NEW only active when
there are exactly two symbols involved in the conflict, one of which
is accessible.

This fixes long-standing test :USE-PACKAGE-CONFLICT-SET in
packages.impure.lisp