Remove bogus logic
A glance at the params files tells you that dynamic-space-start is higher than
any small space, which means that max-dynamic-space-end can only be #xFF..FF
because there is no other space to run into. But the code is doubly bogus
because ever since heap relocation was implemented, DYNAMIC_SPACE_START is 0
when runtime args get parsed, and 0 + the size is NOT the calculated end.
Together these facts make the error unreachable, because no size when added
to 0 wraps around.