Make (LOG 0) always signal division-by-zero
Don't rely on the state of the floating point control register to
trap: the type deriver for LOG assumes that the result type of LOG on
non-negative integer types is bounded below by 0.

This should fix the symptoms of occasional flakes in google's test
infrastructure (but does not address the underlying problem, which
seems to be losing the floating point control word state under
certain, probably signal-handler-related, circumstances).