Adapt LOG derive-type for integer domain lower bound
While (LOG 0) always signals division-by-zero, (LOG 0d0) need not,
depending on the state of the floating point control.  If we are
deriving the type of LOG for a known-integer type with no known lower
bound, use 0 as the lower bound of the real-returning interval rather
than 0d0.

This should fix the flakes in Google's test suite of the LOG
derive-type resulting from the floating point control word somehow,
sometimes not being in the usual state.